refactor: refact project structure
