+++ src/utils/createNewUser.ts
@@ -1,5 +1,5 @@
 import { v4 as uuidv4 } from 'uuid';
-import { User } from '../interface/user';
+import { User } from '../interfaces/user';
 
 export const createUser = (username: string, age: number, hobbies: string[]): User => {
     return {
